+++ R/import_eea_stations.R
@@ -4,69 +4,81 @@
 #' viewer](https://discomap.eea.europa.eu/App/AQViewer/index.html?fqn=Airquality_Dissem.b2g.measurements).
 #' This is useful for binding onto datasets obtained using
 #' [get_eea_parquet_files()] to assign monitoring data with relevant site
-#' information. Metadata can be filtered by providing the `countries` and
-#' `cities` arguments.
+#' information. The user must select a subset of countries using the `countries` argument.
 #'
 #' @inheritParams get_eea_parquet_urls
-#'
-#' @param ... Not currently used; reserved for future functionality.
-#'
-#' @param .cache This function works by installing a CSV to a temporary
-#'   directory and reading it into R. When `.cache` is `TRUE`, this only occurs
-#'   once per R session. When `.cache` is `FALSE`, this CSV is re-downloaded
-#'   every time the function is called.
 #'
 #' @author Jack Davison
 #'
 #' @export
 import_eea_stations <- function(
-  countries = NULL,
-  cities = NA,
-  ...,
-  .cache = TRUE
+  countries = NULL
 ) {
-  td <- tempdir()
-  filename <- "euroaq.dataflowd.csv"
-
-  path <- file.path(td, filename)
-
-  if (file.exists(path) && !.cache) {
-    unlink(path, force = TRUE)
+  if (is.null(countries)) {
+    cli::cli_abort(
+      "Please specify at least one country using {.arg countries}."
+    )
   }
 
-  if (!file.exists(path)) {
-    url <- "https://discomap.eea.europa.eu/App/AQViewer/download?fqn=Airquality_Dissem.b2g.measurements&f=csv"
+  country_codes <- import_eea_countries()
 
-    utils::download.file(url, path, mode = "wb")
-  }
+  countries <- country_codes$country_name[
+    country_codes$country_code %in% countries
+  ]
 
-  out <-
-    suppressWarnings(vroom::vroom(
-      path,
-      show_col_types = FALSE,
-      progress = FALSE,
-      .name_repair = snakecase::to_snake_case
-    )) |>
-    dplyr::tibble()
+  meta <-
+    purrr::map(
+      .x = countries,
+      .f = \(x) {
+        td <- tempdir()
+        zip_filename <- paste0("euroaq_dataflowd_", x, ".zip")
+        dir_filename <- paste0("euroaq_dataflowd_", x)
 
-  if (!is.null(countries)) {
-    spos <-
-      get_eea_country_city_spos(
-        countries = countries,
-        cities = cities,
-        pollutants = NULL,
-        datetime_start = NULL,
-        datetime_end = NULL,
-        dataset = 1,
-        aggregation_type = "hour"
-      )
+        zip_path <- file.path(td, zip_filename)
+        dir_path <- file.path(td, dir_filename)
 
-    out <- dplyr::semi_join(out, spos, dplyr::join_by("sampling_point_id"))
-  }
+        if (file.exists(zip_path)) {
+          unlink(zip_path, recursive = TRUE, force = TRUE)
+        }
+        if (dir.exists(dir_path)) {
+          unlink(dir_path, recursive = TRUE, force = TRUE)
+        }
 
-  if (!.cache) {
-    unlink(path, force = TRUE)
-  }
+        httr2::request(
+          "https://discomap.eea.europa.eu/App/AQViewer/download?fqn=Airquality_Dissem.b2g.measurements&f=csv"
+        ) |>
+          httr2::req_headers(
+            "Content-Type" = "application/json",
+            "Accept" = "application/json, text/*"
+          ) |>
+          httr2::req_body_json(list(
+            Page = 0,
+            RequestFilter = list(
+              Country = list(
+                FieldName = "Country",
+                Values = as.list(x)
+              )
+            )
+          )) |>
+          httr2::req_perform(path = zip_path)
 
-  return(out)
+        utils::unzip(zipfile = zip_path, exdir = dir_path)
+
+        suppressWarnings(
+          vroom::vroom(
+            list.files(
+              dir_path,
+              pattern = "DataExtract.csv",
+              full.names = TRUE
+            ),
+            show_col_types = FALSE,
+            progress = FALSE,
+            .name_repair = snakecase::to_snake_case
+          )
+        )
+      }
+    ) |>
+    dplyr::bind_rows()
+
+  return(meta)
 }